test: mv eip1193 mock to avoid re-running eip1193 tests
